% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AEDAT.collapsed.mad4k}
\alias{AEDAT.collapsed.mad4k}
\title{AEDAT.collapsed.mad4k}
\format{An object of class \code{ExpressionSet} with 4000 rows and 47 columns.}
\usage{
AEDAT.collapsed.mad4k
}
\description{
AEDAT.collapsed.mad4k
}
\keyword{datasets}
