% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AEDAT.collapsed}
\alias{AEDAT.collapsed}
\title{AEDAT.collapsed}
\format{An object of class \code{ExpressionSet} with 19539 rows and 47 columns.}
\usage{
AEDAT.collapsed
}
\description{
AEDAT.collapsed
}
\keyword{datasets}
