% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breast_loi_133p2}
\alias{breast_loi_133p2}
\title{breast_loi_133p2}
\format{An object of class \code{ExpressionSet} with 54675 rows and 87 columns.}
\usage{
breast_loi_133p2
}
\description{
breast_loi_133p2
}
\keyword{datasets}
