% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dm10}
\alias{dm10}
\title{Chemical Pertubations Dataset}
\format{An object of class \code{ExpressionSet} with 9682 rows and 338 columns.}
\usage{
dm10
}
\description{
Chemical Pertubations Dataset
}
\keyword{datasets}
