% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheat.continuous.R
\name{ggheat.make.groups}
\alias{ggheat.make.groups}
\title{\code{ggheat.make.groups} splits eset into list of esets based on phenotype label, do clustering within groups}
\usage{
ggheat.make.groups(eset, labelcol, labelvals, clustFUN, fixRowOrd = TRUE)
}
\arguments{
\item{eset}{expression set}

\item{labelcol}{column name for grouping in pData(eset)}

\item{labelvals}{values to group on (e.g. factor levels of pData(eset)[, labelcol])
clustFUN clustering function for eset, e.g. clust_eset}
}
\description{
\code{ggheat.make.groups} splits eset into list of esets based on phenotype label, do clustering within groups
}
