% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HNSC_htseq_raw_counts_AEvsG1vsG3}
\alias{HNSC_htseq_raw_counts_AEvsG1vsG3}
\title{HNSC_htseq_raw_counts_AEvsG1vsG3}
\format{An object of class \code{ExpressionSet} with 34422 rows and 120 columns.}
\usage{
HNSC_htseq_raw_counts_AEvsG1vsG3
}
\description{
HNSC_htseq_raw_counts_AEvsG1vsG3
}
\keyword{datasets}
