% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HNSC_htseq_raw_counts_toy_AEvsG2}
\alias{HNSC_htseq_raw_counts_toy_AEvsG2}
\title{HNSC_htseq_raw_counts_toy_AEvsG2}
\format{An object of class \code{ExpressionSet} with 54556 rows and 30 columns.}
\usage{
HNSC_htseq_raw_counts_toy_AEvsG2
}
\description{
HNSC_htseq_raw_counts_toy_AEvsG2
}
\keyword{datasets}
