% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HNSC_RNASeq_ES}
\alias{HNSC_RNASeq_ES}
\title{Expression Set Object (TCGA-HNSC)}
\format{An object of class \code{ExpressionSet} with 20263 rows and 566 columns.}
\usage{
HNSC_RNASeq_ES
}
\description{
Expression Set Object (TCGA-HNSC)
}
\keyword{datasets}
