% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HNSC_RNASeq_toy_ES}
\alias{HNSC_RNASeq_toy_ES}
\title{Toy Expression Set Object (TCGA-HNSC)}
\format{An object of class \code{ExpressionSet} with 19844 rows and 25 columns.}
\usage{
HNSC_RNASeq_toy_ES
}
\description{
Toy Expression Set Object (TCGA-HNSC)
}
\keyword{datasets}
