% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{renamedBreastDB}
\alias{renamedBreastDB}
\title{renamedBreastDB}
\format{An object of class \code{ExpressionSet} with 3000 rows and 38 columns.}
\usage{
renamedBreastDB
}
\description{
renamedBreastDB
}
\keyword{datasets}
