% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variationFilter.R
\name{VERBOSE}
\alias{VERBOSE}
\title{subset an expression dataset by extracting genes based on their level of variation}
\usage{
VERBOSE(v, ...)
}
\arguments{
\item{dat}{an expressionSet object}

\item{score}{type of variation measure to use (can be 'mad', 'sd', or 'cv')}

\item{dir}{direction of the gene ranking ('top': highest varying genes; 'bottom': lowest varying genes)}

\item{transform}{transform the data before measuring variation}

\item{ngenes}{number of top (or bottom) genes to extract}

\item{do.plot}{plot the center vs. scale plot with selected genes highlighted}

\item{do.log}{indicates whether to log the "x", "y", or "xy" axes, or "" none}
}
\value{
the filtered expressionSet object
}
\description{
subset an expression dataset by extracting genes based on their level of variation
}
\examples{

# select the top 100 genes by median absolute deviation (mad) and also generate plot

# Use example data, for data set information: ?eSet.brca.100
data(eSet.brca.100)

variationFilter(eSet.brca.100,score='mad',ngenes=50,do.plot=TRUE)

}
