% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zebrafish_cufflinks_counts_fpkm_eSet}
\alias{zebrafish_cufflinks_counts_fpkm_eSet}
\title{zebrafish_cufflinks_counts_fpkm_eSet}
\format{An object of class \code{ExpressionSet} with 24438 rows and 16 columns.}
\usage{
zebrafish_cufflinks_counts_fpkm_eSet
}
\description{
zebrafish_cufflinks_counts_fpkm_eSet
}
\keyword{datasets}
